docs(readme): add gitter badge, remove "é" accents
Added a gitter badge to our list of badges at the top. Also, one of the
badges was "flat-square", and we decided to follow the "square" style
for all our badges, so that's been modified.
Since the é (e acute) character might not be allowed on some services,
we decided to ensure the name always uses a normal "e" in "Code Cafe",
thus the update to our headers.