Fix Heroku deployment (#20)
* chore (config) Add heroku autodeploy

...again

* chore (config) add heroku postinstall

Experimenting with heroku. Maybe adding a postinstall will help.

* fix (config) resolve yarn & host related issues

Yarn was out of sync with package.json, so the yarn file has been
reconstructed.

The server was listening on the localhost hostname, which caused many
much issues with incoming web requests from not localhost. #resolved.

The start script now runs a node script.