config(tests) Add nightwatch & refactor server
Set up and configured nightwatch.js for end-to-end testing. To
facilitate the nightwatch tests, we needed a way to start the server
(preferably with our port of choice), so all the server logic is now in
two files:
- prod-server.js: our production server, automatically listens on a
port.
- server.js: configures all the server logic without actually starting
the server, exporting the express app.
This way our tests can use the same server as production without
horrible hacks.